It is now possible to choose between classic and simple/insecure bruteforce protection method using 'auth.bruteforcingProtectClass' config key (which should be a class name, default to net.rujel.auth.ClassicBruteforceProtection)
